refactor: 예약시간 관리자 API 분리
