+++ src/main/java/roomescape/time/presentation/AdminReservationTimeController.java
@@ -1,0 +1,37 @@
+package roomescape.time.presentation;
+
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.time.application.ReservationTimeService;
+import roomescape.time.presentation.dto.ReservationTimeRequest;
+import roomescape.time.presentation.dto.ReservationTimeResponse;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/admin/times")
+public class AdminReservationTimeController {
+
+    private final ReservationTimeService service;
+
+    @PostMapping
+    public ResponseEntity<ReservationTimeResponse> createReservationTime(@Valid @RequestBody ReservationTimeRequest request) {
+        ReservationTimeResponse response = ReservationTimeResponse.from(
+                service.addReservationTime(request.toEntity())
+        );
+        return ResponseEntity.status(HttpStatus.CREATED).body(response);
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteReservationTime(@PathVariable Long id) {
+        service.deleteReservationTime(id);
+        return ResponseEntity.noContent().build();
+    }
+}
