+++ src/main/java/roomescape/time/presentation/ReservationTimeController.java
@@ -3,21 +3,15 @@ package roomescape.time.presentation;
 import jakarta.validation.Valid;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.time.application.ReservationTimeService;
 import roomescape.time.application.dto.AvailableReservationTimeInfo;
 import roomescape.time.presentation.dto.AvailableReservationTimeRequest;
 import roomescape.time.presentation.dto.AvailableReservationTimeResponse;
-import roomescape.time.presentation.dto.ReservationTimeRequest;
 import roomescape.time.presentation.dto.ReservationTimeResponse;
 
 @RestController
@@ -34,20 +28,6 @@ public class ReservationTimeController {
                 .map(ReservationTimeResponse::from)
                 .toList();
         return ResponseEntity.ok(responses);
-    }
-
-    @PostMapping
-    public ResponseEntity<ReservationTimeResponse> createReservationTime(@Valid @RequestBody ReservationTimeRequest request) {
-        ReservationTimeResponse response = ReservationTimeResponse.from(
-                service.addReservationTime(request.toEntity())
-        );
-        return ResponseEntity.status(HttpStatus.CREATED).body(response);
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteReservationTime(@PathVariable Long id) {
-        service.deleteReservationTime(id);
-        return ResponseEntity.noContent().build();
     }
 
     @GetMapping("/available")
